Create ModelVariables
